evaluate/requirements.txt
scikit-image>=0.22
playwright>=1.40
pdf2image>=1.16
python-pptx>=0.6.23
Pillow>=10.0
beautifulsoup4>=4.12
lxml>=5.0
numpy>=1.26
